Also run tests on PR
This modifies the CI config so that test are also run on PRs to the
repo. This is helpful so that PRs from pull-bot have the tests run
on them.

We take some care to make sure that tests are not run twice
unnecessarily and that docker images are only built on branches not PRs.